files/requirements.txt
GitPython
requests
PyYAML